requirements.txt
Django>=4.2
python-docx>=0.8.11
